Add files via upload
How this program works:

This program will make certain assumption before it begins to run. It will assume that the doctor has already set up
the patient on the app with the required amount of daily medication. This program has been built to be dynamic in it's
representation of the patient's input of the amount of medication they have consumed, via the use of visual graphics.
It has not been built to be adaptive to any changes made by the doctor to the amount of medication that the patient
needs to have on a daily basis, this change will have to be made manually for now. Perhaps through some mechanism it
can be built into the program in the future.

The user will be given two options:

Option 1: Reports generated for each Dose of the medication.

This option is for doctors who would like to administer a stronger level of visibility over their patient's consumption
habits. Particulary in instances of patients who are on very high doses of medication, this option provides a almost
"live" method of keeping track of overdoses and medications missed. Reports will be available when patient administers
the medication during all points of the day.

Option 2: Reports generated at the end of the day.

This option is great for the doctor to administer on a patient who doesn't require as much attention to detail on
their medication consumption. This method will showcase at the end of the day a report of all the medication that
the patient has had/not had and will update the app.

Future additions:

I will continue to work on this project going forward. The aim will be to create an IOS app that does all of the
above. I will also look into certain machine learning algorithms that could allow for the user to just take a picture
of all of the medication they are having from inside of the app. The app will then identify the medications and the
quantity of each of them and generate a report. I am not sure whether this can actually be done, but I am looking
forward to starting the process of figuring this out!